Moved the operatorToString implementation to a source file.
