+++ operators/CMakeLists.txt
@@ -5,6 +5,7 @@ set(
 
 set(
    sources
+   src/operators.cpp
 )
 
 set(
